+++ database/macrostrat/database/utils.py
@@ -101,7 +101,9 @@ def temp_database(conn_string, drop=True, ensure_empty=False):
     """Create a temporary database and tear it down after tests."""
     create_database(conn_string, exists_ok=True, replace=ensure_empty)
     try:
-        yield create_engine(conn_string)
+        engine = create_engine(conn_string)
+        yield engine
+        engine.dispose()
     finally:
         if drop:
             drop_database(conn_string)
@@ -149,7 +151,7 @@ def create_engine(db_conn, **kwargs):
         if isinstance(url, str):
             url = make_url(url)
         # Set the driver to psycopg if not already set
-        if url.drivername != "postgresql+psycopg":
+        if "postgres" in url.drivername:
             url = url.set(drivername="postgresql+psycopg")
 
         return base_create_engine(url, **kwargs)