+++ dinosaur/macrostrat/dinosaur/__init__.py
@@ -4,10 +4,10 @@ from contextlib import contextmanager, redirect_stdout
 from typing import Callable
 
 import docker
-from migra import Migration
-from migra.statements import check_for_drop
+from results.dbdiff import Migration
+from results.dbdiff.statements import check_for_drop
+from results.schemainspect import get_inspector
 from rich import print
-from schemainspect import get_inspector
 
 from macrostrat.database import Database
 from macrostrat.database.utils import connection_args, run_sql, temp_database
@@ -103,7 +103,7 @@ def _target_db(
 
     log.debug("Creating migration target")
     with temp_database(url) as engine:
-        database = Database(url)
+        database = Database(engine)
         with redirect_stdout(redirect):
             initializer(database)
         yield engine